Moved touch and ls into subdirectories
Fixed Makefile to account for per-utility subdirectories
